@@ -2,8 +2,6 @@ package de.metzgore.rbtvschedule;
 
 import android.app.Application;
 import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
 
 public class RBTVScheduleApp extends Application {
 
@@ -15,17 +13,8 @@ public class RBTVScheduleApp extends Application {
         mAppContext = this;
     }
 
+    //TODO dagger
     public static Context getAppContext() {
         return mAppContext.getApplicationContext();
-    }
-
-    public static boolean hasNetwork() {
-            return mAppContext.checkIfHasNetwork();
-    }
-
-    public boolean checkIfHasNetwork() {
-        ConnectivityManager cm = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo networkInfo = cm.getActiveNetworkInfo();
-        return networkInfo != null && networkInfo.isConnected();
     }
 }
